Extract AttachmentRedirectUrlUpdater#update! into worker
It is better practice when queuing a asynchronous worker to allow the
worker to look up the data it needs when the work is done, rather than
provide it with the data when the job is enqueued. This commit
extracts the logic from the AttachmentRedirectUrlUpdater service
listener into a new worker. Currently we call that worker
synchronously because we are passing an instance of AttachmentData to
the worker and the definition of this class could conceivably change
if there is a long delay between queuing and the work being done. This
will be fixed in a subsequent commit.

The tests for AssetManagerAttachmentRedirectUrlUpdateWorker have been
moved from the service listener.
