+++ app/services/service_listeners/attachment_redirect_url_updater.rb
@@ -11,21 +11,7 @@ module ServiceListeners
 
     def update!
       return unless attachment_data.present?
-      redirect_url = nil
-      if attachment_data.unpublished?
-        redirect_url = attachment_data.unpublished_edition.unpublishing.document_url
-      end
-      enqueue_job(attachment_data.file, redirect_url)
-      if attachment_data.pdf?
-        enqueue_job(attachment_data.file.thumbnail, redirect_url)
-      end
-    end
-
-  private
-
-    def enqueue_job(uploader, redirect_url)
-      legacy_url_path = uploader.asset_manager_path
-      AssetManagerUpdateAssetWorker.perform_async(legacy_url_path, redirect_url: redirect_url)
+      AssetManagerAttachmentRedirectUrlUpdateWorker.new.perform(attachment_data)
     end
   end
 end
