+++ app/workers/asset_manager_attachment_redirect_url_update_worker.rb
@@ -1,0 +1,19 @@
+class AssetManagerAttachmentRedirectUrlUpdateWorker < WorkerBase
+  def perform(attachment_data)
+    redirect_url = nil
+    if attachment_data.unpublished?
+      redirect_url = attachment_data.unpublished_edition.unpublishing.document_url
+    end
+    enqueue_job(attachment_data.file, redirect_url)
+    if attachment_data.pdf?
+      enqueue_job(attachment_data.file.thumbnail, redirect_url)
+    end
+  end
+
+private
+
+  def enqueue_job(uploader, redirect_url)
+    legacy_url_path = uploader.asset_manager_path
+    AssetManagerUpdateAssetWorker.perform_async(legacy_url_path, redirect_url: redirect_url)
+  end
+end
